+++ lib/src/command/generate_command.dart
@@ -1,6 +1,7 @@
 import 'package:slidy/slidy.dart';
 import 'package:slidy/src/command/sub_command/generate_bloc.dart';
 import 'package:slidy/src/command/sub_command/generate_controller.dart';
+import 'package:slidy/src/command/sub_command/generate_model_sub_command.dart';
 import 'package:slidy/src/command/sub_command/generate_service_sub_command.dart';
 import 'package:slidy/src/command/sub_command/generate_test_sub_command.dart';
 
@@ -27,6 +28,8 @@ class GenerateCommand extends CommandBase {
     addSubcommand(GenerateServiceAbbrSubCommand());
     addSubcommand(GenerateTestSubCommand());
     addSubcommand(GenerateTestAbbrSubCommand());
+    addSubcommand(GenerateModelSubCommand());
+    addSubcommand(GenerateModelAbbrSubCommand());
   }
 }